move test logic out to unit_tests location
